services/notifications_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pika==1.3.2
redis==5.0.1
python-decouple==3.8
celery==5.3.4
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
